pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "base58>=2.1.1",
    "bcrypt>=4.3.0",
    "ecdsa>=0.19.1",
    "google-auth>=2.40.3",
    "psycopg2-binary>=2.9.10",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "pymysql-pool>=0.5.0",
    "pytelegrambotapi>=4.29.1",
    "requests>=2.32.5",
    "tornado>=6.5.2",
    "urllib3>=2.5.0",
    "valr-python>=0.2.7",
]
